moved the func to get repo & owner to constructStacker
